:hammer: Change Scene's ReturnValue associetedType
**Detail**
This commit change `Scene`'s ReturnValue gained through `rewind method` was Optional to Non Optional.
Because, it is inconvenient because it is necessary to check nil when receiving RetuneValue.

**Breaking Change**
initialize `TransitionRequest` have to change this.

- before

```swift
let request = HogeRequest() { ret in // result is optional
   guard let result = result else { return }
   // process using result
}
```
- after

```swift
let request = HogeRequest() { ret in // result is non-optional
   // process using result
}
```